# perf: GetField reduce allocations