One less alloc/op, but numbers can matter at scale.


Before:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: Apple M1 Pro
BenchmarkTxGetField-10    	 2294287	       582.5 ns/op	     448 B/op	       9 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/corazawaf	2.374s

BenchmarkCRSSimplePOST-10    	     902	   1276337 ns/op	  338771 B/op	    7220 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/testing/coreruleset	2.054s

```
After:
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: Apple M1 Pro
BenchmarkTxGetField-10    	 3306369	       355.6 ns/op	     400 B/op	       8 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/corazawaf	2.051s

BenchmarkCRSSimplePOST-10    	     874	   1262635 ns/op	  320807 B/op	    6418 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/testing/coreruleset	2.034s
```